refactor: 선착순 퀴즈 모달 내 퀴즈 제출 로직 -> 퀴즈 스텝으로 이동
