+++ packages/user/src/components/shared/modal/fcfs/index.tsx
@@ -1,15 +1,12 @@
 import { QueryErrorResetBoundary } from '@tanstack/react-query';
-import { Suspense, useEffect } from 'react';
+import { Suspense } from 'react';
 import { ErrorBoundary } from 'react-error-boundary';
 import Modal, { ModalProps } from 'src/components/common/Modal.tsx';
 import PendingStep from 'src/components/shared/modal/PendingStep.tsx';
-import useSubmitFCFSQuiz, { SubmitFCFSQuizResponse } from 'src/hooks/query/useSubmitFCFSQuiz.ts';
 import useFunnel from 'src/hooks/useFunnel.ts';
 import ErrorStep from './ErrorStep.tsx';
 import QuizStep from './QuizStep.tsx';
 import ResultStep from './ResultStep.tsx';
-
-export type ResultStepType = ReturnType<typeof getResultStepFromStatus>;
 
 export default function FCFSModal(props: ModalProps) {
 	const [Funnel, setStep] = useFunnel(
@@ -26,18 +23,6 @@ export default function FCFSModal(props: ModalProps) {
 		},
 	);
 
-	const { isPending, mutate: submitAnswer } = useSubmitFCFSQuiz();
-
-	useEffect(() => {
-		if (isPending) setStep('pending');
-	}, [isPending]);
-
-	const handleSubmit = (answer: number) =>
-		submitAnswer(
-			{ answer },
-			{ onSuccess: (response) => setStep(getResultStepFromStatus(response)) },
-		);
-
 	return (
 		<Modal {...props}>
 			<div className="flex h-full w-full items-center justify-center p-[100px]">
@@ -47,7 +32,7 @@ export default function FCFSModal(props: ModalProps) {
 							{({ reset }) => (
 								<ErrorBoundary onReset={reset} FallbackComponent={ErrorStep}>
 									<Suspense fallback={<PendingStep>선착순 퀴즈 불러오는 중...</PendingStep>}>
-										<QuizStep onSelect={handleSubmit} />
+										<QuizStep onStepChange={setStep} />
 									</Suspense>
 								</ErrorBoundary>
 							)}
@@ -74,17 +59,4 @@ export default function FCFSModal(props: ModalProps) {
 			</div>
 		</Modal>
 	);
-}
-
-function getResultStepFromStatus({ status }: SubmitFCFSQuizResponse) {
-	switch (status) {
-		case 'END':
-			return 'end';
-		case 'PARTICIPATED':
-			return 'already-done';
-		case 'WRONG':
-			return 'wrong-answer';
-		default:
-			return 'correct-answer';
-	}
 }
